requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
openai==0.27.8
pydantic==2.0.4
requests==2.31.0